MANY Cron (#2)
* wip: Things are messy but almost working

* wip: Refactor to the general proposed solution

It should be easy to add new type of tasks to be scheduled

* Code refactoring

Split the code in modules

* wip: Add persistent storage support

I'm logging the full `send` response to the DB

* wip: Refactoring and better error handling

* wip: Add documentation and better error handling

* wip: Update tokio and tracing-subscriber crates

Enable the 'signal' feature in tokio-cron-scheduler

* wip: Delayed application shutdown

Wait 5secs when a shutdown signal is received to allow some time for the
scheduled tasks to finish. We want to make sure we avoid the case where
a transaction is made but the response is not logged in the persistent
storage.

* wip: Remove useless comments

* wip: Add unit tests

* wip: Use CBOR for Params, better error handling

- Task parameters are now CBOR string instead of plain JSON.
- Task-related errors are now logged in the persistent storage
- Improved the persistent storage key generation

* wip: fmt

* wip: Use public https git URLs

* wip: Use tempfile instead of tempdir

tempdir is a deprecated crate

* Address comments from code review

- Add Cargo.lock
- Add .gitignore
- Add rust-toolchain.yml
- Removed persistent storage
- Code refactoring

There are still some TODO left, but nothing "critical". The more
pressing one is the scheduler shutdown handler.

* Add issue url

Co-authored-by: Hans Larsen <hans@larsen.online>

Co-authored-by: Hans Larsen <hans@larsen.online>